make pageMode a computed property, and move vuex stuff to its own dir
